update to newer version

+++ src/main/java/net/dirtcraft/discord/discordlink/Commands/Discord/StopServer.java
@@ -1,9 +1,10 @@
 package net.dirtcraft.discord.discordlink.Commands.Discord;
 
 import cpw.mods.fml.common.FMLCommonHandler;
-import net.dirtcraft.discord.discordlink.API.GameChat;
+import net.dirtcraft.discord.discordlink.API.Channels;
 import net.dirtcraft.discord.discordlink.API.MessageSource;
 import net.dirtcraft.discord.discordlink.Commands.DiscordCommandExecutor;
+import net.dirtcraft.discord.discordlink.Commands.Sources.DiscordResponder;
 
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
@@ -18,8 +19,9 @@ public class StopServer implements DiscordCommandExecutor {
     @Override
     public void execute(MessageSource source, String command, List<String> args) {
         try {
-            GameChat.sendEmbed("Discord-Link Reboot", "Attempting to reboot the server.");
-            source.getMessage().delete().queue(s->{},e->{});
+            if (!source.isPrivateMessage()) source.getMessage().delete().queue(s->{},e->{});
+            else source.sendCommandResponse("Forced Reboot Scheduled.", "Attempting to reboot the server.");
+            Channels.getDefaultChat().sendMessage(source, "Forced Reboot Scheduled.", "Attempting to reboot the server.");
             Thread.sleep(555);
         } catch (Throwable ignored){
