+++ src/main/java/net/dirtcraft/discord/discordlink/Commands/DiscordCommandManager.java
@@ -1,6 +1,6 @@
 package net.dirtcraft.discord.discordlink.Commands;
 
-import net.dirtcraft.discord.discordlink.API.GameChat;
+import net.dirtcraft.discord.discordlink.API.Channels;
 import net.dirtcraft.discord.discordlink.API.MessageSource;
 import net.dirtcraft.discord.discordlink.API.Roles;
 import net.dirtcraft.discord.discordlink.Commands.Discord.*;
@@ -60,30 +60,68 @@ public class DiscordCommandManager extends DiscordCommandTree {
                 .setRequiredRoles(Roles.STAFF)
                 .build();
 
+        DiscordCommand ranks = DiscordCommand.builder()
+                .setDescription("Reveals a players ranks.")
+                .setCommandExecutor(new Ranks())
+                .setRequiredRoles(Roles.VERIFIED)
+                .build();
+
+        DiscordCommand kits = DiscordCommand.builder()
+                .setDescription("Reveals a players kits.")
+                .setCommandExecutor(new Kits())
+                .setRequiredRoles(Roles.VERIFIED)
+                .build();
+
         DiscordCommand unverify = DiscordCommand.builder()
                 .setDescription("Unverifies your account.")
                 .setCommandExecutor(new Unlink())
                 .build();
 
-        register(list, "list");
-        register(halt, "halt");
+        DiscordCommand prefix = DiscordCommand.builder()
+                .setDescription("Sets prefixes")
+                .setCommandUsage("<title>")
+                .setRequiredRoles(Roles.STAFF)
+                .setCommandExecutor(new Prefix())
+                .build();
+
+        DiscordCommand version = DiscordCommand.builder()
+                .setDescription("Shows the current version")
+                .setRequiredRoles(Roles.DIRTY)
+                .setCommandExecutor(new Version())
+                .build();
+
+        DiscordCommand logs = DiscordCommand.builder()
+                .setDescription("Shows latest logs")
+                .setRequiredRoles(Roles.DIRTY)
+                .setCommandExecutor(new Logs())
+                .build();
+
+        register(list, "list", "players");
         register(stop, "stop");
+        register(halt, "halt");
         register(seen, "seen");
         register(unstuck, "unstuck", "spawn");
         register(username, "username");
         register(discord, "discord");
+        register(ranks, "ranks", "groups", "parents");
+        //register(sync, "sync");
         register(unverify, "unverify", "unlink");
+        //register(notify, "notify");
+        register(prefix, "prefix");
+        register(kits, "kits");
+        register(version, "version", "info");
+        register(logs, "logs");
     }
 
     public void process(MessageSource member, String args){
         try {
-            String[] command = args == null || defaultAliases.contains(args)? new String[0] : args.toLowerCase().split(" ");
+            String[] command = args == null || defaultAliases.contains(args)? new String[0] : args.split(" ");
             execute(member, null, new ArrayList<>(Arrays.asList(command)));
         } catch (Exception e){
             String message = e.getMessage() != null? e.getMessage() : "an error occurred while executing the command.";
             Utility.sendCommandError(member, message);
         } finally {
-            member.getMessage().delete().queue();
+            if (!member.isPrivateMessage()) member.getMessage().delete().queue();
         }
     }
 
@@ -95,8 +133,8 @@ public class DiscordCommandManager extends DiscordCommandTree {
             if (!cmd.hasPermission(member)) return;
             String title = pre + alias + " " + cmd.getUsage();
             embed.addField(title, cmd.getDescription(), false);
+            embed.setFooter("Requested By: " + member.getUser().getAsTag(), member.getUser().getAvatarUrl());
         });
-        GameChat.sendMessage(embed.build());
+        member.sendCommandResponse(embed.build());
     }
-
 }
